Admin:Dates/Bookings: Holy crap! Almost there!
Woohoo! Finished most of the backend work to display the booking admin
panel.

Got the email html to be spewed properly for easy copy/paste to GMail.
Refactored a bit of AdminDisplay's reminder email code to make this
easier.

I still need to incorporate the Dates/Timeframes into the actual
bookings script. This is a HUGE chunk of my time being saved.

Next steps:
* Test out the backend on the real system.
* Let users view/edit their own dates
* Get feedback and adjust the interface
* Then, start working on Templates
* After templates, ready for full on automation work.