Admin: Dates/Bookings: Partial checkin
Getting the booking script with both the booking-script friendly dates
and timeframes.

The freakin' load times are horrid! I need to get xdebug and profile
where the time is getting sucked up. Probably all those stupid for loops
with dates and timeframes.

For some reason, not able to get the NEWLINE to seperate the bookings
script dates. October dates and November dates should be separated by a
NEWLINE delimiter the script will understand.