Category: Partial reworking of submit
Figuring out "submit" with javascript.
Removed the form's action which was pointing to remove venues.
Need to either redirect to a page or just pop up  a div and handle via
ajax.

This is just a checkin of what I've done so far. Not working code.
Breaking change for removing venues!!!!