BDD: Fix navigating to dates page
Fixed navigating to dates page. I was clicking on the list item of the
menu instead of the anchor of the menu. Apparently, to selenium, there's
that much of a difference.

Tried cleaning up some login code. Sometimes it gets a little wonky and
puts the password in the username field. But, even with the explicit
wait on that field, it would still do it. So I just added a 1 second
wait which seems to do the trick.