Bookings: Bulk Pause/Start implemented
Bulk actions of pausing and start on multiple venues has been
implemented.

User can choose one, a few, or all venues to act on.

Removed the ajax-y coolness of the start/pause/resume buttons. But I
think it's more intuitive and simple this way.

Need to organize the rows by active/paused.
Need to color them to visually show which are active and paused.