Bookings: Fixed filtering. Adding tests
In my javascript, I was looking for an integer for the "paused" field,
but the integer was being represented by a string in the venues.

Also, I was looking for venue.id instead of venue.venue_id. This caused
the input names to be "value_undefined" instead of the venues id.

Beginning to implement BDD scenarios.