Frequency: Initial work
Hacking my way around.

Tried to do a pop-over to update, but it's way too much work to get
things done quickly. Just doing the classic form submit to a different
page.

Starting work on a different page.

Need to update the javascript to redirect the form to the frequency edit
page.